igniteflow_-the-generative-venture-studio/backend/requirements.txt
websockets
google-cloud-aiplatform
google-auth
